+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/overlay/renderer/ScreenNameOverlayRenderer.kt
@@ -1,4 +1,4 @@
-package com.donglab.screennameviewer.overlay
+package com.donglab.screennameviewer.overlay.renderer
 
 import android.app.Activity
 import android.content.Context
@@ -7,6 +7,8 @@ import android.widget.FrameLayout
 import android.widget.LinearLayout
 import android.widget.TextView
 import com.donglab.screennameviewer.config.ScreenNameOverlayConfig
+import com.donglab.screennameviewer.overlay.builder.OverlayLayoutBuilder
+import com.donglab.screennameviewer.overlay.builder.StyledTextViewBuilder
 import com.donglab.screennameviewer.util.dp
 import com.donglab.screennameviewer.util.getStatusBarHeight
 import java.lang.ref.WeakReference
